pyproject.toml
[tool.poetry]
name = "nvd-severity"
version = "0.1.0"
description = ""
authors = ["Nikita Pivkin <nikita.pivkin@smartforce.io>"]
readme = "README.md"

[tool.poetry.scripts]
nvd-severity = "nvd_severity.main:run"

[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.64.1"
aiolimiter = "^1.0.0"
backoff = "^2.2.1"
python-dotenv = "^1.0.0"
aiofiles = "^23.1.0"
aiohttp = "^3.8.4"
aiodns = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
